SearchWritersをfeaturesに移行 #33
