+++ src/app/page.tsx
@@ -2,8 +2,7 @@ import { getDatabase } from "@/utils/notion";
 import { Props as ArticleListItemProps } from "@/components/ArticleList/ArticleListItem";
 import { ArticleList } from "@/components/ArticleList";
 import { SearchTags, Tag } from "@/components/SearchTags";
-import { Props as WriterProps } from "@/components/Writer";
-import { SearchWriters } from "@/components/SearchWriters";
+import { SearchWriters } from "@/features/SearchWriters";
 
 export default async function Home({
   searchParams,
@@ -57,39 +56,15 @@ export default async function Home({
     return uniqueTags as Tag[];
   };
 
-  const fetchAllWriters = async () => {
-    const databaseId = process.env.NOTION_DATABASE_ID;
-    const articleDb = await getDatabase(databaseId);
-
-    const results = articleDb
-      .filter((article: any) => {
-        return article.properties.Publish.checkbox === true;
-      })
-      .map((article: any) => {
-        return {
-          name: article.properties.Writer.created_by.name,
-          image: article.properties.Writer.created_by.avatar_url,
-        };
-      });
-
-    const uniqueWriters = results.filter(
-      (writer: any, index: number, self: any) =>
-        self.findIndex((w: any) => w.name === writer.name) === index
-    );
-
-    return uniqueWriters as WriterProps[];
-  };
-
   const articles = await fetchArticles();
   const tags = await fetchAllTags();
-  const writers = await fetchAllWriters();
 
   return (
     <main className="my-90 flex justify-center gap-90">
       <ArticleList articles={articles} />
       <div className="flex flex-col gap-60">
         <SearchTags tags={tags} />
-        <SearchWriters writers={writers} />
+        <SearchWriters />
       </div>
     </main>
   );
