+++ src/features/SearchWriters/index.tsx
@@ -1,0 +1,8 @@
+import React from "react";
+import { SearchWritersPresentation } from "./presentations/";
+import { useWriters } from "./hooks";
+
+export const SearchWriters = async () => {
+  const writers = await useWriters();
+  return <SearchWritersPresentation writers={writers} />;
+};
